+++ example/android/app/src/main/kotlin/com/example/gap_widgets/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.gap_plus
+package com.example.gap_widgets
 
 import io.flutter.embedding.android.FlutterActivity