+++ lib/gap_plus.dart
@@ -1,2 +1,0 @@
-export 'package:gap_plus/src/widgets/gap.dart';
-export 'package:gap_plus/src/widgets/sliver_gap.dart';