+++ lib/src/rendering/rendering.dart
@@ -1,0 +1,2 @@
+export 'render_gap.dart';
+export 'render_sliver_gap.dart';